+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased][main]
 
+## [0.2.0] - 2021-08-12
+
 - Add Log action for $profile_update event with $succeeded and $failed statuses during reset password process
 - Add Risk action for $profile_update event with $attempted status and Log action for $profile_update event with $succeeded and $failed statuses
 - Add Log action for $password_reset_request event with $succeeded and $failed statuses
@@ -12,4 +14,5 @@
 - Initial release
 
 [main]: https://github.com/castle/castle_devise/compare/v0.1.0...HEAD
+[0.2.0]: https://github.com/castle/castle_devise/compare/v0.2.0...v0.1.0
 [0.1.0]: https://github.com/castle/castle_devise/releases/tag/v0.1.0